backend/pyproject.toml
[project]
name = "django-oauth-backend"
version = "0.1.0"
description = "Django backend with DRF, Sentry, and Google OAuth2"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "django>=4.2.7,<5.0",
    "djangorestframework>=3.14.0,<4.0",
    "django-cors-headers>=4.3.1,<5.0",
    "sentry-sdk>=1.38.0,<2.0",
    "social-auth-app-django>=5.4.0,<6.0",
    "requests>=2.31.0,<3.0",
    "python-decouple>=3.8,<4.0",
    "gunicorn>=21.2.0,<22.0",
    "whitenoise>=6.6.0,<7.0",
    "PyJWT>=2.8.0,<3.0",
    "cryptography>=41.0.0,<42.0",
    "mysqlclient>=2.1.0,<3.0",
    "dj-database-url>=2.0.0,<3.0",
    "boto3>=1.34.0,<2.0",
    "django-storages>=1.14.0,<2.0",
    "django-mcp-server>=0.5.6",
    "django-stubs[compatible-mypy]>=5.2.2",
    "djangorestframework-stubs[compatible-mypy]>=3.16.1",
    "matplotlib>=3.7.0,<4.0",
    "pytz>=2023.3,<2024.0",
    "slack-sdk>=3.36.0",
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "users"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mysql-connector-python>=8.0.0,<9.0",
    "types-boto3>=1.40.1",
    "django-stubs>=5.1.3",
    "djangorestframework-stubs>=3.15.1",
    "mypy>=1.14.1",
    "types-pytz>=2023.3.0",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = false
disable_error_code = ["var-annotated", "no-any-return", "unused-ignore"]

[[tool.mypy.overrides]]
module = [
    "django.*",
    "djangorestframework.*",
    "social_auth_app_django.*",
    "boto3.*",
    "botocore.*",
]
ignore_missing_imports = true 
